+++ ESFramework/Foundation/NSURL+ESAdditions.h
@@ -25,6 +25,21 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (NSURL *)URLByAddingQueryDictionary:(NSDictionary<NSString *, id> *)queryDictionary;
 
+/**
+ * e.g. "https://itunes.apple.com/app/id12345678?mt=8"
+ */
++ (NSURL *)appLinkWithIdentifier:(NSInteger)appIdentifier;
+
+/**
+ * e.g. "itms-apps://itunes.apple.com/app/id12345678"
+ */
++ (NSURL *)appStoreLinkWithIdentifier:(NSInteger)appIdentifier;
+
+/**
+ * e.g. "itms-apps://itunes.apple.com/WebObjects/MZStore.woa/wa/viewContentsUserReviews?type=Purple+Software&id=12345678"
+ */
++ (NSURL *)appStoreReviewLinkWithIdentifier:(NSInteger)appIdentifier;
+
 @end
 
 NS_ASSUME_NONNULL_END